시연 커밋

+++ src/main/java/controller/Usercontroller.java
@@ -1,7 +1,5 @@
 package controller;
 
-import java.util.HashMap;
-import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
@@ -9,10 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 import exception.LoginException;
 import logic.User;
@@ -48,15 +44,6 @@ public class Usercontroller {
 		return mav;
 	}
 	
-	@RequestMapping("user/idcheck")
-    @ResponseBody
-    public Map<String, Integer> idcheck(@RequestBody String id) {
-        int count = 0;
-        Map<String, Integer> map = new HashMap<String, Integer>();
-        count = userService.idcheck(id);
-        map.put("cnt", count);
-        return map;
-    }
 	@RequestMapping(value="user/login", method=RequestMethod.GET)
 	public ModelAndView loginForm() {
 		ModelAndView mav = new ModelAndView();
@@ -78,6 +65,7 @@ public class Usercontroller {
 				mav.addObject("dbUser", dbUser);
 				mav.setViewName("main");
 				session.setAttribute("loginUser", dbUser);
+				
 			} else {
 				bindingResult.reject("error.login.password");
 				mav.getModel().putAll(bindingResult.getModel());
@@ -104,48 +92,24 @@ public class Usercontroller {
 	}
 	
 	@RequestMapping("user/mypage")//view를 띄워주는 메소드
-	public ModelAndView mypage() {
+	public ModelAndView mypage(HttpSession session) {
 	ModelAndView mav = new ModelAndView();
-	mav.addObject(new User());
+	User loginUser = (User)session.getAttribute("loginUser");//수정시에 xml에 selectOne(값을 하나만 받아옴. 여러개를 받을 수 없음)을 쓰면서 xml조건문에 id != null을 쓰기 때문에 처음에 이 페이지를 띄울 때는 기존에 등록해 놓은 loginUser의 정보로 조회를 함. 그리고 수정한 후에 띄울 때는 수정한 정보를 띄움.-재한이오빠가 알려줌ㅋㅋ
+	User dbuser = userService.dbuser(loginUser.getId());
+	mav.addObject("dbuser", dbuser);
 	return mav;
 	}
 	
-	/*@RequestMapping(value="user/update", method=RequestMethod.POST)
-	public ModelAndView update(@Valid User user, BindingResult bindingResult, HttpSession session, HttpServletRequest request) {
-	ModelAndView mav = new ModelAndView("user/mypage");
-	if(bindingResult.hasErrors()) {
-		mav.getModel().putAll(bindingResult.getModel());
-		return mav;
-	}
-	User loginUser = (User) session.getAttribute("loginUser");
-	User dbUser = userService.getUser(user.getId());
-	if(loginUser.getPw().equals("admin")) {
-		if(!loginUser.getPw().equals(user.getPw())) {
-			throw new LoginException("관리자 비밀번호가 올바르지 않습니다.", "mypage.zips?id="+user.getId()); 
-		}
-	} else {
-		if(!user.getPw().equals(dbUser.getPw())) {
-			throw new LoginException("비밀번호가 틀리죠^^?", "mypage.user?id="+user.getId());
-		}
-	}
-	try {
-		userService.updateUser(user);
-		mav.setViewName("redirect:/user/mypage.zips");
-	} catch (Exception e) {
-		e.printStackTrace();
-	}
-		return mav;
-	}*/
-	
 	@RequestMapping("user/update")
-	public ModelAndView update(@Valid User user, BindingResult bindingResult, HttpServletRequest request) {
+	public ModelAndView update(@Valid User user, BindingResult bindingResult, HttpServletRequest request, String id) {
 		ModelAndView mav = new ModelAndView("user/mypage");
 		if(bindingResult.hasErrors()) {
 			mav.getModel().putAll(bindingResult.getModel());
 		return mav;
 		}
 		userService.updateUser(user, request);
-		mav.setViewName("redirect:/user/mypage.zips");
+		mav.addObject("dbuser", user);
+		mav.setViewName("redirect:mypage.zips?id="+user.getId());//setViewName은 .jsp가 붙음
 		return mav;
 	}
 	
@@ -159,16 +123,19 @@ public class Usercontroller {
 	
 	@RequestMapping(value="user/delete", method=RequestMethod.POST)
 	public ModelAndView delete(@Valid User user, BindingResult bindingResult, HttpSession session) {
-		ModelAndView mav = new ModelAndView("user/delete.zips?id="+user.getId());
+		ModelAndView mav = new ModelAndView("user/delete");
 		if(bindingResult.hasErrors()) {
 			mav.getModel().putAll(bindingResult.getModel());
-			mav.setViewName("user/delete.zips?id=" + user.getId());
+			mav.setViewName("user/delete");
 			return mav;
 		}
 		User dbUser = userService.getUser(user.getId());
 		User loginUser = (User)session.getAttribute("loginUser");
 		if (!user.getPw().equals(dbUser.getPw())) {
 			throw new LoginException("비밀번호가 틀립니다.", "delete.zips?id="+user.getId());
+		}
+		if (user.getPw() == null || user.getPw().isEmpty()) {
+			throw new LoginException("비밀번호를 입력해 주세요.", "delete.zips?id="+user.getId());
 		}
 		try {
 			userService.delete(user.getId());
